backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain
langchain-openai
langchain-community
langchain-groq
openai
langfuse
chromadb
sentence-transformers
openpyxl
twilio
httpx
# Socket.IO for real-time updates
python-socketio
python-engineio
aiofiles
# Database
sqlalchemy
pymysql
# explicitly pin numpy to avoid incompatibility with pandas
numpy<1.24